pyproject.toml
[project]
name = "monday"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Giovanni", email = "42703789+gbronca@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "python-dotenv>=1.0.1",
    "ruff>=0.7.1",
]
